Avoid using datetime.__new__ directly
The datetime.__new__() method should return an instance with properties
we can work with when inheriting from the datetime class, but it seems
that isn't the case. Any datetime method that even references the
private attributes set in __new__ are completely inaccessible, which was
causing some very odd behavior.

Most of the changes in this commit serve to remove dependency on those
built-in methods and attributes that should work, but don't given we're
assigning new values to _year, _month, etc. Perhaps there is a more
elegant way to avoid this problem altogether, but discovering such a
solution is time-prohibitive… especially when you consider how little
time is required to copy/paste code from the datetime class.